Find more 'round'-related variables of types int32 and uint32 and change them to 64-bit

+++ integrationTests/vm/iele/vmAgar_test.go
@@ -22,7 +22,7 @@ func TestDeployAgarioContract(t *testing.T) {
 	senderAddressBytes := []byte("12345678901234567890123456789012")
 	senderNonce := uint64(11)
 	senderBalance := big.NewInt(100000000)
-	round := uint32(444)
+	round := uint64(444)
 	gasPrice := uint64(1)
 	gasLimit := uint64(1000000)
 
@@ -57,7 +57,7 @@ func TestAgarioContractTopUpShouldWork(t *testing.T) {
 	senderAddressBytes := []byte("12345678901234567890123456789012")
 	senderNonce := uint64(11)
 	senderBalance := big.NewInt(100000000)
-	round := uint32(444)
+	round := uint64(444)
 	gasPrice := uint64(1)
 	gasLimit := uint64(1000000)
 
@@ -115,7 +115,7 @@ func TestAgarioContractTopUpAnfWithdrawShouldWork(t *testing.T) {
 	senderAddressBytes := []byte("12345678901234567890123456789012")
 	senderNonce := uint64(11)
 	senderBalance := big.NewInt(100000000)
-	round := uint32(444)
+	round := uint64(444)
 	gasPrice := uint64(1)
 	gasLimit := uint64(1000000)
